refactor : 고객의 대출 종류에 따른 처리 로직 수정

+++ BankManagerConsoleApp/BankManagerConsoleApp/Customer.swift
@@ -9,5 +9,5 @@ import Foundation
 
 struct Customer {
     let bankNumberTicket: Int
-    let bankingType: Int
+    let bankingType: Double
 }
